+++ src/components/resources/imageImports.js
@@ -1,0 +1,23 @@
+// imports/exports all images used in resource page
+import { ReactComponent as Drop } from "../../images/drop.svg";
+import { ReactComponent as Cover } from "../../images/cover.svg";
+import { ReactComponent as HoldOn } from "../../images/holdon.svg";
+import { ReactComponent as CaneDrop } from "../../images/cane-drop.svg";
+import { ReactComponent as CaneCover } from "../../images/cane-cover.svg";
+import { ReactComponent as CaneHoldOn } from "../../images/cane-holdon.svg";
+import { ReactComponent as WalkerLock } from "../../images/walker-lock.svg";
+import { ReactComponent as WalkerCover } from "../../images/walker-cover.svg";
+import { ReactComponent as WalkerHoldOn } from "../../images/walker-holdon.svg";
+import { ReactComponent as WheelchairBrake } from "../../images/wheelchair-brake.svg";
+import { ReactComponent as WheelchairCover } from "../../images/wheelchair-cover.svg";
+import { ReactComponent as WheelchairHoldOn } from "../../images/wheelchair-holdon.svg";
+import { ReactComponent as BulletPlus } from "../../images/bullet-plus.svg";
+import { ReactComponent as BulletSolid } from "../../images/bullet-solid.svg";
+import { ReactComponent as TriangleUp } from "../../images/triangle-up.svg";
+import { ReactComponent as STEPLogo } from "../../images/STEP-logo.svg";
+import FEMALogo from "../../images/fema-logo.jpg";
+import RedCrossLogo from "../../images/red-cross-logo.jpg";
+import SalvationArmyLogo from "../../images/salvation-army-logo.jpg";
+import EarthquakeShelterLogo from "../../images/earthquake-shelter-logo.jpg";
+
+export { Drop, Cover, HoldOn, CaneDrop, CaneCover, CaneHoldOn, WalkerLock, WalkerCover, WalkerHoldOn, WheelchairBrake, WheelchairCover, WheelchairHoldOn, BulletPlus, BulletSolid, TriangleUp, STEPLogo, FEMALogo, RedCrossLogo, SalvationArmyLogo, EarthquakeShelterLogo };
